Completed exercises 1.21-23 product, filter-in and list-index
In order to keep the exercises interesting, I implemented product
using Racket's built-in append-map (because it's only natural to
implement the function this way).

filter-in was written the long way, and as a bonus I implemented
complement (like Clojure) and defined remove-in in terms of
filter-in for fun.

Lastly, list-index has two conceptual parts: the function that
returns the Integer index of the first element matching a
predicate, and the part that returns the specified output
(Int | Bool). Treating these as separate problems allows
us to write a more general index-finding function that returns
a simple Integer. Conceptually, there are two functions:

  index-of: (Any -> Bool) x List -> Int, or 'f'
  is-found: Int -> (Int | Bool), or 'g'

list-index is therefore the composition of these functions (g o f)

Signed-off-by: Christian Romney <cromney@pointslope.com>