Consistently raise error on invalid environment
By consistently we mean that an invalid environment will
*eventually* raise an error in any of the representations,
although the error may be raised at different logical
times in different representations.

In the datatype representation, for instance, the error is
raised when attempting to extend-env. In the closure (or
procedure) representation, the error isn't detected until
apply-env is called with the invalid environment.

Signed-off-by: Christian Romney <cromney@pointslope.com>