Completed exercises 2.9 amd 2.10 has-binding? and extend-env*
Only the closure (procedure) environment representation lacks
support for the has-binding? function. This is because
applying an empty-env to a variable throws an eopl:error,
which the eopl language does not have support for handling.

All representations support the empty-env? predicate.

Only the association list representation supports extend-env*.

Modified the run-for-all macro to accept any number of
functions.

Signed-off-by: Christian Romney <cromney@pointslope.com>