Completed exercises 2.13, 2.14 (fix procedure implementation)
Temporarily skipped 2.12

This commit extends the procedural implementation to support
both empty-env? and has-binding? by changing the internal
representation of empty-env and extend-env to return a dispatch
function.

The dispatch function switches on a "method name" and returns
another procedure which can still make use of the free variables
from the constructor function. With this change, we can know
if we have a binding for the search variable because we maintain
our reference to both the var and the saved environment.

Signed-off-by: Christian Romney <cromney@pointslope.com>