Exercise 2.16 / 2.17* Lambda Calculus variation
Exercise 2.16 requires removal of parens around bound var
in expressions.

Exercise 2.17 asks for two more representations. Since I don't really
need the Scheme practice, I chose to sort of combine the two into
a single, new shorthand representation that uses λ instead of lambda,
drops the parentheses as requested, and inserts a "." to separate the
bound var and the body. This bears some resemblance to the
representations found in academic literature on the Lambda Calculus.

Signed-off-by: Christian Romney <cromney@pointslope.com>