Completed 2.20 (with tests) and added tests for Exercise 2.19
Created a new representation of binary trees which behaves
identically to the version that can't move 'up' but
which keeps track of its parent node (at insertion time).

I didn't follow the book's "hint" to use a reversed list,
since it seems much more straightforward to add a parent
element to the bintree representation instead.

Bintree ::= () | (Int Bintree Bintree Bintree)

In this representation, the leaf node is identical to the
previous representation, but the interior node gains a new
Bintree at the end of the sequence, representing its parent.

This representation is inefficient with respect to space,
because the child node is given a complete copy of its
parent node. This cannot be avoided until we have a way
to hold a reference (instead of merely a value).

Signed-off-by: Christian Romney <cromney@pointslope.com>